+++ src/main/java/org/example/service/impl/DeptServiceImpl.java
@@ -1,10 +1,13 @@
 package org.example.service.impl;
 
+import org.example.excetion.ClassStudentException;
 import org.example.mapper.DeptMapper;
+import org.example.mapper.EmpMapper;
 import org.example.pojo.Dept;
 import org.example.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -14,12 +17,18 @@ import java.util.List;
 public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
     @Override
     public List<Dept> findAll() {
         return deptMapper.findAll();
     }
+    @Transactional
     @Override
     public void deleteById(Integer id) {
+        if (empMapper.countById(id)>0){
+            throw new ClassStudentException("此部门下还有员工!");
+        }
         deptMapper.deleteById(id);
     }
     @Override
